updated from tag to posttag naming and fetching
