+++ json-server.py
@@ -8,7 +8,8 @@ from views import get_all_posts
 from views import create_user, login_user
 from views import get_single_post
 from views import grabCategoryList, addCategory
-from views import getTagList, get_post_tags, addTag
+from views import getTagList, addTag
+from views import get_post_tags
 from views import get_all_comments, get_post_comments, create_comment
 
 
@@ -48,12 +49,16 @@ class JSONServer(HandleRequests):
         
         elif resource == "tag":
             if url['pk'] != 0:
-                response_body = get_post_tags(url)
-                return self.response(response_body, status.HTTP_200_SUCCESS.value)
+                pass
             else:
                 response_body = getTagList()
                 return self.response(response_body, status.HTTP_200_SUCCESS.value)
-
+            
+        elif resource == "posttag":
+            if url['pk'] != 0:
+                response_body = get_post_tags(url)
+                return self.response(response_body, status.HTTP_200_SUCCESS.value)
+            
         else:
             return self.response("", status.HTTP_404_CLIENT_ERROR_RESOURCE_NOT_FOUND.value)
         
@@ -61,6 +66,7 @@ class JSONServer(HandleRequests):
 
     def do_PUT(self):
         pass
+
 
     def do_DELETE(self):
         pass
@@ -108,7 +114,6 @@ class JSONServer(HandleRequests):
                 return self.response("", status.HTTP_204_SUCCESS_NO_RESPONSE_BODY.value)
             else:
                 return self.response("Requested Resource not found", status.HTTP_404_CLIENT_ERROR_RESOURCE_NOT_FOUND.value)
-
 
 #APPARENTLY NO ONE CARES ABOUT THIS
 def main():
