+++ views/__init__.py
@@ -2,6 +2,7 @@ from .user import create_user, login_user
 from .my_posts_view import get_all_posts
 from .post import get_single_post
 from .category import grabCategoryList, addCategory
-from .tag import getTagList, get_post_tags, addTag
+from .tag import getTagList, addTag
+from .postTags import get_post_tags
 from .comment import get_post_comments, get_all_comments, create_comment
